Use proper null values in addBook function
Use the nullable values from the sql package to properly handle null values for
series, subtitle, edition and purchased_date values when adding new books to the
database.

The advantage of doing it this way is not only that it ensures NULLs are handled
properly and the database maintains proper form. It also has the happy effect of
removing a bodge in my code which sought to handle NULL values for series_id and
resulted in significant code duplication.